+++ packages/kitchensink/service-hash/feature-home.js
@@ -1,0 +1,10 @@
+const homeRouteHandler = async (request) => {
+  const hash = request.getContext('hash');
+  return hash.encode('foo');
+};
+
+module.exports = {
+  name: 'home',
+  hook: '$FASTIFY_GET',
+  handler: [{ url: '/', handler: homeRouteHandler }],
+};