+++ packages/service-hash/src/hooks.js
@@ -1,0 +1,2 @@
+const SERVICE_NAME = `hash`;
+module.exports = { SERVICE_NAME };