+++ CHANGELOG.md
@@ -23,6 +23,8 @@
   - Update to `setContext` API
 - service-jwt
   - Remove Babel
+- service-hash
+  - Remove Babel
 - service-fastify
   - Fixes `$REGISTER_ROUTE` hook so to be synchronous and report injections in the boot report
   - Remove babel and the build step
@@ -33,6 +35,7 @@
   - Added examples for `@forrestjs/hooks`
   - Added examples for `@forrestjs/service-fastify`
   - Added examples for `@forrestjs/service-jwt`
+  - Added examples for `@forrestjs/service-hash`
   - Added `app01` examples for a complex app
 
 ## 3.20.2