+++ packages/kitchensink/CHANGELOG.md
@@ -5,4 +5,5 @@
 - Added examples for `@forrestjs/hooks`
 - Added examples for `@forrestjs/service-fastify`
 - Added examples for `@forrestjs/service-jwt`
+- Added examples for `@forrestjs/service-hash`
 - Added `app01` examples for a complex app