+++ packages/service-hash/CHANGELOG.md
@@ -1,2 +1,5 @@
 # Change Log
 
+## v1.21.0
+
+- Remove Babel